+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## [1.0.2] - 2025-07-02
+
+### Fixed
+- Fixed transaction response handling
+
 ## [1.0.1] - 2025-07-02
 
 ### Added
@@ -16,6 +21,7 @@
 ### Fixed
 - Fixed service support checking logic in JNAPService
 - Corrected version numbers for service capabilities
+- Fixed transaction response handling
 
 ## [1.0.0] - 2025-06-26